api/requirements.txt
flask>=3.0
flask-cors
requests
pymupdf
python-pptx
python-docx
numpy
python-dotenv
APScheduler>=3.10
